requirements.txt
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
chromedriver-binary==86.0.4240.22.0
cycler==0.10.0
idna==2.10
japanize-matplotlib==1.1.3
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.11
